% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{Create package}
\usage{
create_package(directory, package_name)
}
\arguments{
\item{directory}{string. Folder directory where repositories are stored.}

\item{package_name}{string. Name of package to be created.}
}
\description{
Create package
}
